backend/requirements.txt
fastapi>=0.95.0
uvicorn[standard]>=0.23.0
SQLAlchemy>=1.4
alembic>=1.10
# Use pydantic v1 for this scaffold (v2 has incompatible changes to types/helpers)
pydantic>=1.10,<2
python-dotenv>=1.0
loguru>=0.7
